feat: Create Register Panel + logic
