--- packages/ui/components/Panels/Modal/ModalContext.ts
@@ -1,11 +1,0 @@
-import { createContext } from "react";
-
-const ModalContext = createContext({
-	labelProps: "",
-	multipleLabels: [""],
-	inputType: "",
-	multipleInputs: [""],
-	state: { errorMessages: { inputType: null } },
-});
-
-export default ModalContext;
