+++ packages/ui/components/Panels/Modal/ModalLabel.tsx
@@ -1,6 +1,13 @@
 import { FormLabel } from "@chakra-ui/react";
 import { useContext } from "react";
-import ModalContext from "./ModalContext";
+import ModalContext from "./state/ModalContext";
+
+const LabelStyle = {
+	color: "#D6BCFA",
+	position: "absolute",
+	top: "-5px",
+	paddingLeft: "5px",
+} as React.CSSProperties;
 
 export const Label = ({ labelProps }: { labelProps?: string }) => {
 	const { multipleLabels } = useContext(ModalContext);
@@ -9,18 +16,7 @@ export const Label = ({ labelProps }: { labelProps?: string }) => {
 			{labelProps ? (
 				<FormLabel
 					sx={{ position: "relative", margin: "0.25rem " }}
-					requiredIndicator={
-						<span
-							style={{
-								color: "#D6BCFA",
-								position: "absolute",
-								top: "-5px",
-								paddingLeft: "5px",
-							}}
-						>
-							＊
-						</span>
-					}
+					requiredIndicator={<span style={LabelStyle}>＊</span>}
 				>
 					{labelProps}
 				</FormLabel>
