+++ packages/ui/components/Panels/Modal/state/ModalContext.ts
@@ -1,0 +1,12 @@
+import { createContext } from "react";
+import { Action } from "./ModalReducer";
+
+const ModalContext = createContext({
+	state: {
+		isTyping: false,
+		errorMessages: {},
+	},
+	dispatch: (arg: Action) => {},
+});
+
+export default ModalContext;
