+++ packages/ui/components/Panels/Modal/state/definitions.d.ts
@@ -1,0 +1,13 @@
+export type ModalState = {
+	errorMessages:
+		| {
+				inputType: string;
+		  }
+		| {};
+	labelProps: string;
+	multipleLabels: string[];
+	inputType: string | null;
+	multipleInputs: string[];
+	test: string;
+	isTyping: boolean;
+};
